+++ gakumas-tools/components/Simulator/Simulator.js
@@ -316,7 +316,7 @@ export default function Simulator() {
             href="https://github.com/surisuririsu/gakumas-tools/blob/master/gakumas-tools/simulator/CHANGELOG.md"
             target="_blank"
           >
-            {t("lastUpdated")}: 2026-02-18
+            {t("lastUpdated")}: 2026-02-19
           </a>
         </div>
       </div>