+++ packages/gakumas-engine/engine/CardManager.js
@@ -121,6 +121,7 @@ export default class CardManager extends EngineComponent {
     state[S.holidayCardsUsed] = 0;
     state[S.onigiriCardsUsed] = 0;
     state[S.koeteCardsUsed] = 0;
+    state[S.kyakkouCardsUsed] = -1;
   }
 
   changeIdol(state) {